requirements.txt
numpy
matplotlib
wordcloud
pillow